requirements.txt
polars 
jax 
optax
pandas 
numpy
tqdm
xgboost
lightgbm
scikit-learn
matplotlib
sklego
mlforecast
holidays 
neuralforecast
kedro-datasets 
